+++ spec/restful_model_collection_spec.rb
@@ -1,5 +1,5 @@
-describe Nylas::RestfulModelCollection do
-  let(:api) { Nylas::API.new(app_id, app_secret, access_token) }
+describe NylasDashboardV2SDK::RestfulModelCollection do
+  let(:api) { NylasDashboardV2SDK::API.new(app_id, app_secret, access_token) }
   let(:app_id) { 'ABC' }
   let(:app_secret) { '123' }
   let(:access_token) { 'UXXMOCJW-BKSLPCFI-UQAQFWLO' }
@@ -86,7 +86,7 @@ describe Nylas::RestfulModelCollection do
                    :body => '{"message": "Could not verify access credential.",'+
                             '"type": "invalid_request_error"}')
 
-      expect { api.threads.count }.to raise_error(Nylas::AccessDenied)
+      expect { api.threads.count }.to raise_error(NylasDashboardV2SDK::AccessDenied)
     end
   end